Module slosh_test::docs

Structs

DOC_REGEX 🔒

DocStringSection 🔒

EXEMPTIONS

SloshDoc 🔒

Enums

DocError 🔒

Namespace 🔒

Constants

DESCRIPTION 🔒

EXAMPLE 🔒

SECTION 🔒

USAGE 🔒

Traits

AsMd

Functions

add_builtins

build_all_slosh_forms_listing_chapter 🔒

build_doc 🔒

build_each_docs_section_chapter 🔒

build_symbols_list 🔒

capitalize_first 🔒

doc_map 🔒

get_docs_by_section 🔒

Each doc has a tag in its Section: definition by convention that logically groups functions. Using a HashMap store the section tags as keys and add all slosh docs from to a vector as a value corresponding to its section.

get_exemptions 🔒

get_globals_sorted 🔒

insert_nil_section 🔒

insert_section 🔒

make_file 🔒

name_for_all_page 🔒

name_for_section_page 🔒

symbol_and_capitalized_symbol 🔒

Type Aliases

DocResult 🔒